feat: add cover page form field component
